+++ src/main/java/de/thomas_oster/liblasercut/drivers/Lasersaur.java
@@ -163,8 +163,8 @@ public class Lasersaur extends LaserCutter {
           break;
         case SETPROPERTY:
           PowerSpeedFocusFrequencyProperty p = (PowerSpeedFocusFrequencyProperty) cmd.getProperty();
-          setPower(out, p.getPower());
-          setSpeed(out, p.getSpeed());
+          setPower(out, (int) p.getPower());
+          setSpeed(out, (int) p.getSpeed());
           break;
       }
     }
@@ -202,7 +202,7 @@ public class Lasersaur extends LaserCutter {
     boolean dirRight = true;
     Point rasterStart = rp.getRasterStart();
     PowerSpeedFocusProperty prop = (PowerSpeedFocusProperty) rp.getLaserProperty();
-    setSpeed(out, prop.getSpeed());
+    setSpeed(out, (int) prop.getSpeed());
     ByteArrayList bytes = new ByteArrayList(rp.getRasterWidth());
     for (int line = 0; line < rp.getRasterHeight(); line++) {
       Point lineStart = rasterStart.clone();
@@ -227,7 +227,7 @@ public class Lasersaur extends LaserCutter {
               if (old == 0) {
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               } else {
-                setPower(out, prop.getPower() * (0xFF & old) / 255);
+                setPower(out, (int) prop.getPower() * (0xFF & old) / 255);
                 line(out, lineStart.x + pix - 1, lineStart.y, resolution);
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               }
@@ -235,7 +235,7 @@ public class Lasersaur extends LaserCutter {
             }
           }
           //last point is also not "white"
-          setPower(out, prop.getPower() * (0xFF & bytes.get(bytes.size() - 1)) / 255);
+          setPower(out, (int) prop.getPower() * (0xFF & bytes.get(bytes.size() - 1)) / 255);
           line(out, lineStart.x + bytes.size() - 1, lineStart.y, resolution);
         } else {
           //move to the last nonempty point of the line
@@ -246,7 +246,7 @@ public class Lasersaur extends LaserCutter {
               if (old == 0) {
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               } else {
-                setPower(out, prop.getPower() * (0xFF & old) / 255);
+                setPower(out, (int) prop.getPower() * (0xFF & old) / 255);
                 line(out, lineStart.x + pix + 1, lineStart.y, resolution);
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               }
@@ -254,7 +254,7 @@ public class Lasersaur extends LaserCutter {
             }
           }
           //last point is also not "white"
-          setPower(out, prop.getPower() * (0xFF & bytes.get(0)) / 255);
+          setPower(out, (int) prop.getPower() * (0xFF & bytes.get(0)) / 255);
           line(out, lineStart.x, lineStart.y, resolution);
         }
       }
@@ -269,8 +269,8 @@ public class Lasersaur extends LaserCutter {
     boolean dirRight = true;
     Point rasterStart = rp.getRasterStart();
     PowerSpeedFocusProperty prop = (PowerSpeedFocusProperty) rp.getLaserProperty();
-    setSpeed(out, prop.getSpeed());
-    setPower(out, prop.getPower());
+    setSpeed(out, (int) prop.getSpeed());
+    setPower(out, (int) prop.getPower());
     for (int line = 0; line < rp.getRasterHeight(); line++) {
       Point lineStart = rasterStart.clone();
       lineStart.y += line;
@@ -304,7 +304,7 @@ public class Lasersaur extends LaserCutter {
               if (old == 0) {
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               } else {
-                setPower(out, prop.getPower() * (0xFF & old) / 255);
+                setPower(out, (int) prop.getPower() * (0xFF & old) / 255);
                 line(out, lineStart.x + pix - 1, lineStart.y, resolution);
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               }
@@ -312,7 +312,7 @@ public class Lasersaur extends LaserCutter {
             }
           }
           //last point is also not "white"
-          setPower(out, prop.getPower() * (0xFF & bytes.get(bytes.size() - 1)) / 255);
+          setPower(out, (int) prop.getPower() * (0xFF & bytes.get(bytes.size() - 1)) / 255);
           line(out, lineStart.x + bytes.size() - 1, lineStart.y, resolution);
           //add some space to the right
           move(out, Math.min((int) Util.mm2px(bedWidth, resolution), (int) (lineStart.x + bytes.size() - 1 + Util.mm2px(this.addSpacePerRasterLine, resolution))), lineStart.y, resolution);
@@ -327,7 +327,7 @@ public class Lasersaur extends LaserCutter {
               if (old == 0) {
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               } else {
-                setPower(out, prop.getPower() * (0xFF & old) / 255);
+                setPower(out, (int) prop.getPower() * (0xFF & old) / 255);
                 line(out, lineStart.x + pix + 1, lineStart.y, resolution);
                 move(out, lineStart.x + pix, lineStart.y, resolution);
               }
@@ -335,7 +335,7 @@ public class Lasersaur extends LaserCutter {
             }
           }
           //last point is also not "white"
-          setPower(out, prop.getPower() * (0xFF & bytes.get(0)) / 255);
+          setPower(out, (int) prop.getPower() * (0xFF & bytes.get(0)) / 255);
           line(out, lineStart.x, lineStart.y, resolution);
           //add some space to the left
           move(out, Math.max(0, (int) (lineStart.x - Util.mm2px(this.addSpacePerRasterLine, resolution))), lineStart.y, resolution);
